used sepolgen to create a standard policy for zfs
Ran sepolgen --init /sbin/zfs

Tweaked it to include zpool executable as well, and ran the included
zfs.sh with the --update parameter to iron out the remaining avc
denials. I'm not sure if the result is necessarily the best set of
rules; they work so I'm accepting them as a zeroth-order approximation.